edge_compute_service/requirements.txt
redis
python-dotenv
llama-index-core
llama-index-llms-ollama
llama-index-embeddings-ollama
llama-index-vector-stores-weaviate
weaviate-client
pypdf